fix 64bit build error
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-2414
Reviewed-on: https://android.intel.com:443/306580

+++ audio_conversion/src/Resampler.hpp
@@ -52,7 +52,7 @@ public:
     android::status_t resampleFrames(const void *src,
                                      void *dst,
                                      const size_t inFrames,
-                                     uint32_t *outFrames);
+                                     size_t *outFrames);
 
     /**
      * Configures the resampler.
